// milkglass-sweets/src/data/seasonal-enum.ts
export enum SeasonalName {
  Birthday = 'Birthday',
  Spring = 'Spring',
  Summer = 'Summer',
  Fall = 'Fall',
  Winter = 'Winter',
  Easter = 'Easter',
  FourthOfJuly = '4th of July',
  Halloween = 'Halloween',
  Thanksgiving = 'Thanksgiving',
  Christmas = 'Christmas',
  NewYears = 'New Years',
}